chore: consolidate big number helpers
