+++ src/components/Layout/Header/NavBar/MobileNavBar.tsx
@@ -5,7 +5,7 @@ import { useTranslate } from 'react-polyglot'
 import { Link as ReactRouterLink, useLocation } from 'react-router-dom'
 import { routes } from 'Routes/RoutesCommon'
 import { usePlugins } from 'context/PluginProvider/PluginProvider'
-import { bnOrZero } from 'lib/investor/investor-foxy'
+import { bnOrZero } from 'lib/bignumber/bignumber'
 
 export const MobileNavBar = () => {
   const { routes: pluginRoutes } = usePlugins()
