+++ src/lib/investor/investor-foxy/index.ts
@@ -1,3 +1,2 @@
 export * from './api'
 export * from './constants'
-export * from './utils'
