--- src/lib/investor/investor-foxy/utils/bignumber.ts
@@ -1,10 +1,0 @@
-import BigNumber from 'bignumber.js'
-
-export type BN = BigNumber
-
-export const bn = (n: BigNumber.Value, base = 10): BN => new BigNumber(n, base)
-
-export const bnOrZero = (n: BigNumber.Value | null | undefined): BN => {
-  const value = bn(n || 0)
-  return value.isNaN() ? bn(0) : value
-}
