+++ src/lib/investor/investor-idle/idlecli.ts
@@ -4,8 +4,8 @@ import type { NativeAdapterArgs } from '@shapeshiftoss/hdwallet-native'
 import { NativeHDWallet } from '@shapeshiftoss/hdwallet-native'
 import * as unchained from '@shapeshiftoss/unchained-client'
 import dotenv from 'dotenv'
+import { bn } from 'lib/bignumber/bignumber'
 import { IdleInvestor } from 'lib/investor/investor-idle/IdleInvestor'
-import { bn } from 'lib/investor/investor-idle/utils'
 
 dotenv.config()
