+++ src/lib/investor/investor-idle/index.ts
@@ -1,5 +1,4 @@
 export * from 'lib/investor/investor-idle/constants'
-export * from 'lib/investor/investor-idle/utils'
 export * from 'lib/investor/investor-idle/IdleSdk'
 export * from 'lib/investor/investor-idle/IdleInvestor'
 export * from 'lib/investor/investor-idle/IdleOpportunity'
