--- src/lib/investor/investor-idle/utils/index.ts
@@ -1,1 +1,0 @@
-export * from 'lib/investor/investor-idle/utils/bignumber'
