+++ src/lib/investor/investor-yearn/index.ts
@@ -1,4 +1,3 @@
 export * from './constants'
-export * from './utils'
 export * from './YearnInvestor'
 export * from './YearnOpportunity'
