--- src/lib/investor/investor-yearn/utils/index.ts
@@ -1,1 +1,0 @@
-export * from './bignumber'
