+++ src/pages/Defi/components/EligibleCarousel.tsx
@@ -5,8 +5,7 @@ import { NavLink } from 'react-router-dom'
 import type { CardProps } from 'components/Card/Card'
 import { Card } from 'components/Card/Card'
 import { Carousel } from 'components/Carousel/Carousel'
-import { bn } from 'lib/bignumber/bignumber'
-import { bnOrZero } from 'lib/investor/investor-foxy'
+import { bn, bnOrZero } from 'lib/bignumber/bignumber'
 import { selectAggregatedEarnUserStakingEligibleOpportunities } from 'state/slices/selectors'
 import { useAppSelector } from 'state/store'
